+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [1.19.0] - 2023-08-03
 ### Added
+- Analytics API for getting posts [#382](https://github.com/rokwire/groups-building-block/issues/382)
 - Prepare for deployment in OpenShift [#379](https://github.com/rokwire/groups-building-block/issues/379)
 
 ## [1.18.2] - 2023-05-03